requirements.txt
dj-database-url==0.5.0
Django==2.1.3
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
pytz==2018.7
whitenoise==4.1.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
